added server.js and made the project functional.
