+++ .gitignore
@@ -6,3 +6,4 @@ node_modules
 .env.test.local
 .env.production.local
 package-lock.json
+instructions.txt
